refactoring get/set regulating terminal
