+++ java/src/main/java/com/powsybl/dataframe/network/adders/GeneratorDataframeAdder.java
@@ -13,6 +13,7 @@ import com.powsybl.dataframe.update.IntSeries;
 import com.powsybl.dataframe.update.StringSeries;
 import com.powsybl.dataframe.update.UpdatingDataframe;
 import com.powsybl.iidm.network.*;
+import com.powsybl.python.network.NetworkUtil;
 
 import java.util.Collections;
 import java.util.List;
@@ -92,7 +93,8 @@ public class GeneratorDataframeAdder extends AbstractSimpleAdder {
             applyIfPresent(ratedS, row, adder::setRatedS);
             applyBooleanIfPresent(voltageRegulatorOn, row, adder::setVoltageRegulatorOn);
             applyIfPresent(energySource, row, EnergySource.class, adder::setEnergySource);
-            applyIfPresent(regulatingElements, row, elementId -> Util.setRegulatingTerminal(adder::setRegulatingTerminal, network, elementId));
+            applyIfPresent(regulatingElements, row, elementId -> NetworkUtil
+                    .setRegulatingTerminal(adder::setRegulatingTerminal, network, elementId));
             return adder;
         }
     }
